applying php-cs-fixer
