+++ ProcessMaker/Http/Controllers/TaskController.php
@@ -28,11 +28,13 @@ use ProcessMaker\Models\UserResourceView;
 use ProcessMaker\Nayra\Contracts\Bpmn\ScriptTaskInterface;
 use ProcessMaker\Traits\HasControllerAddons;
 use ProcessMaker\Traits\SearchAutocompleteTrait;
+use ProcessMaker\Traits\TaskControllerIndexMethods;
 
 class TaskController extends Controller
 {
     use SearchAutocompleteTrait;
     use HasControllerAddons;
+    use TaskControllerIndexMethods;
 
     private static $dueLabels = [
         'open' => 'Due',
